+++ db/migrate/20240313195149_add_private_flag_to_setups.rb
@@ -1,0 +1,5 @@
+class AddPrivateFlagToSetups < ActiveRecord::Migration[7.1]
+  def change
+    add_column :setups, :private, :boolean
+  end
+end