+++ db/migrate/20240313201312_make_private_flag_on_setups_mandatory.rb
@@ -1,0 +1,5 @@
+class MakePrivateFlagOnSetupsMandatory < ActiveRecord::Migration[7.1]
+  def change
+    change_column :setups, :private, :boolean, null: false
+  end
+end