+++ db/migrate/20240313214005_set_default_value_for_private_in_setups.rb
@@ -1,0 +1,5 @@
+class SetDefaultValueForPrivateInSetups < ActiveRecord::Migration[7.1]
+  def change
+    change_column_null :setups, :private, false, false
+  end
+end