+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.1].define(version: 2024_03_12_002323) do
+ActiveRecord::Schema[7.1].define(version: 2024_03_13_214005) do
   create_schema "heroku_ext"
 
   # These are extensions that must be enabled in order to support this database
@@ -221,6 +221,7 @@ ActiveRecord::Schema[7.1].define(version: 2024_03_12_002323) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.integer "user_id"
+    t.boolean "private", null: false
     t.index ["name", "user_id"], name: "index_setups_on_name_and_user_id", unique: true
   end